Add cell selection user interaction
Create a file to hold bindings, and created binding to inform SUD if
cell was clicked.

Small restructuring of SUD, now stores the current board instance as a
property.  Also created method on SUD to react to this binding by
making calls to board (to update data), and then calling render methods
to refresh the display.

Additional bindings (e.g. for entering/clearing cell values) should
follow this same approach.