Add user value entry using menu bar
User can now add/erase values from cells using the menu bar. Created
bindings on menu, which calls a SUD method to update the board state and
call the appropriate render method.

Need to add a differentiating class for 'locked' cell values that are
starting values (and the user shouldn't be able to edit).